+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # WIP
 
 - msec argument to squuid (#95, #97)
+- lookup refs in JS API `pull`, `pull_many`, `entity` (#94, thx [Matt Senior](https://github.com/mattsenior))
 
 # 0.11.5